[#217,#219] Introduce proper handling of package revision strings
This commit introduces a get_package_filename function to build.py that
generates the full package version, including a package revision string, which
we have not been supplying until this point. This package revision string
contains the distro codename for deb-based distros, and the EL major version
for rpm-based distros.

A `package_revision` entry is now supported in versions.json, to specify the
base package revision, to which the distro-specific suffix will be appended. If
the entry is not present, "0" will be assumed.

Despite the fact that we have not been supplying package revision strings
until now, "2" was chosen as the `package_revision` starting point for all
current packages, as our rpm packages have contained a false package revision
of "1" in their filenames.